+++ lib/super_image.dart
@@ -1,26 +1,27 @@
 // ignore_for_file: unnecessary_import
 library super_image;
-// --------------------------------------------------------------------------
-import 'package:flutter/material.dart';
-import 'package:super_image/super_image/b_super_image_box.dart';
-import 'package:super_image/super_image/c_image_switcher.dart';
-import 'package:widget_fader/widget_fader.dart';
-import 'package:bldrs_theme/bldrs_theme.dart';
-import 'package:flutter/services.dart';
-import 'package:super_image/helpers/helpers.dart';
 import 'dart:typed_data';
 import 'dart:ui' as ui;
-import 'package:super_image/helpers/ui_checker.dart';
+
+import 'package:bldrs_theme/bldrs_theme.dart';
+// --------------------------------------------------------------------------
+import 'package:filers/filers.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:mapper/mapper.dart';
+import 'package:super_image/src/b_super_image_box.dart';
+import 'package:widget_fader/widget_fader.dart';
+
+import 'src/c_image_switcher.dart';
 // --------------------------------------------------------------------------
 /// SUPER IMAGE
-part './super_image/a_super_image.dart';
-// --------------------
-/// INFINITY LOADING BOX
-part './super_image/x_infinity_loading_box.dart';
-// --------------------
-/// LOCAL ASSET CHECKER
-part './super_image/x_local_asset_checker.dart';
-// --------------------
+part 'src/a_super_image.dart';
 /// CACHELESS IMAGE
-part './super_image/x_cacheless_image.dart';
+part 'src/x_cacheless_image.dart';
+/// FUTURE IMAGE
+part 'src/x_future_bytes_image.dart';
+/// INFINITY LOADING BOX
+part 'src/x_infinity_loading_box.dart';
+/// LOCAL ASSET CHECKER
+part 'src/x_local_asset_checker.dart';
 // --------------------------------------------------------------------------